Update dependencies and bump version to v0.6.1

+++ crates/ai00-core/src/sampler/mod.rs
@@ -7,16 +7,16 @@ mod radix;
 
 pub trait Sampler {
     /// Initialize the sampler state.
-    fn init(&mut self, model_tokens: &[u16]);
+    fn init(&mut self, model_tokens: &[u32]);
     /// Update the raw model output.
     fn transform(&self, output: &mut [f32]);
     /// Select one token from the distribution, and also update the state.
-    fn sample(&mut self, probs: &[f32]) -> u16;
+    fn sample(&mut self, probs: &[f32]) -> u32;
 }
 
 pub trait Formatter {
     /// Update the raw model output.
     fn transform(&self, output: &mut [f32]);
     /// Update the internal state after a token is chosen. Return if the state machine is halt.
-    fn update(&mut self, token: u16) -> bool;
+    fn update(&mut self, token: u32) -> bool;
 }
